+++ lib/sefaz.rb
@@ -2,9 +2,7 @@
 
 require 'savon'
 
-require_relative "sefaz/version"
-
-module SEFAZ
-  class Error < StandardError; end
-  # Your code goes here...
-end
+require_relative "sefaz/utils/version"
+require_relative "sefaz/utils/connection"
+require_relative "sefaz/utils/wsdl"
+require_relative "sefaz/nfe"